Clean up jobs a little
