+++ metadeploy/api/cci_configs.py
@@ -1,0 +1,35 @@
+from cumulusci.core.config import YamlProjectConfig
+
+
+class MetadeployProjectConfig(YamlProjectConfig):
+    def __init__(self, *args, repo_root=None, **kwargs):  # pragma: nocover
+        self._repo_root = repo_root
+        super().__init__(*args, **kwargs)
+
+    @property
+    def repo_root(self):  # pragma: nocover
+        return self._repo_root
+
+    @property
+    def config_project_local_path(self):  # pragma: nocover
+        return
+
+    @property
+    def repo_name(self):  # pragma: nocover
+        return
+
+    @property
+    def repo_url(self):  # pragma: nocover
+        return
+
+    @property
+    def repo_owner(self):  # pragma: nocover
+        return
+
+    @property
+    def repo_branch(self):  # pragma: nocover
+        return
+
+    @property
+    def repo_commit(self):  # pragma: nocover
+        return
